+++ src/odometry.cpp
@@ -7,7 +7,7 @@
 #include <cmath>
 
 namespace okapi {
-OdomState::OdomState(const float ix, const float iy, const float itheta)
+OdomState::OdomState(const double ix, const double iy, const double itheta)
   : x(ix), y(iy), theta(itheta) {
 }
 
@@ -16,15 +16,15 @@ OdomState::OdomState() : x(0), y(0), theta(0) {
 
 OdomState::~OdomState() = default;
 
-OdometryParams::OdometryParams(const ChassisModelParams &iparams, const float iscale,
-                               const float iturnScale)
+OdometryParams::OdometryParams(const ChassisModelParams &iparams, const double iscale,
+                               const double iturnScale)
   : model(iparams.make()), scale(iscale), turnScale(iturnScale) {
 }
 
 OdometryParams::~OdometryParams() = default;
 
-Odometry::Odometry(const ChassisModelParams &imodelParams, const float iscale,
-                   const float iturnScale)
+Odometry::Odometry(const ChassisModelParams &imodelParams, const double iscale,
+                   const double iturnScale)
   : model(imodelParams.make()), scale(iscale), turnScale(iturnScale), lastTicks{0, 0}, mm(0) {
 }
 
@@ -38,7 +38,7 @@ Odometry::Odometry(const OdometryParams &iparams)
 
 Odometry::~Odometry() = default;
 
-void Odometry::setScales(const float iscale, const float iturnScale) {
+void Odometry::setScales(const double iscale, const double iturnScale) {
   scale = iscale;
   turnScale = iturnScale;
 }
@@ -50,10 +50,10 @@ void Odometry::loop() {
   while (true) {
     newTicks = model->getSensorVals();
     tickDiff = newTicks - lastTicks;
-    mm = (static_cast<float>(tickDiff[1] + tickDiff[0]) / 2.0) * scale;
+    mm = (static_cast<double>(tickDiff[1] + tickDiff[0]) / 2.0) * scale;
     lastTicks = newTicks;
 
-    state.theta += (static_cast<float>(tickDiff[1] - tickDiff[0]) / 2.0) * turnScale;
+    state.theta += (static_cast<double>(tickDiff[1] - tickDiff[0]) / 2.0) * turnScale;
     if (state.theta > 180)
       state.theta -= 360;
     else if (state.theta < -180)
